Move container resolver to its own package
